+ Implement Basic CRUD for UEmployment, ULicense, UEduction, UCertification on UI and DashboardController

+++ src/main/webapp/assets/javascripts/javawork/user_profile_overview.js
@@ -4,23 +4,106 @@ var currentItem;
 function bindCurrentCertification(certificationId){
 	currentItem = certificationId;
 	
-	$("#uConferringOrganization").val($("#item-conferringOrganization-"+ certificationId+" h5").text());
+	$("#certification-conferringOrganization").val($("#certification-conferringOrganization-"+ currentItem+" h5").text());
 	
-	$("#uProfessionalOrganization").val($("#item-professionalCertificate-"+ certificationId).text());
+	$("#certification-professionalOrganization").val($("#certification-professionalCertificate-"+ currentItem).text());
 	
-	$("#uDateAwarded").val($("#item-dateAwarded-"+ certificationId).text());
+	$("#certification-dateAwarded").val($("#certification-dateAwarded-"+ currentItem).text());
 	
-	$("#uCertificateNumber").val($("#item-certificateNumber-"+ certificationId).text());
+	$("#certification-certificateNumber").val($("#certification-certificateNumber-"+ currentItem).text());
 	
-	$("#uDescription").val($("#item-description-"+ certificationId).text());
+	$("#certification-description").val($("#certification-description-"+ currentItem).text());
 }
 
 function updateUCertification() {
-	alert("123");
-	$("#UpdateUCertification").attr("action", "dashboard/" + currentItem + "/updateUCertification.jv");
-	$("#UpdateUCertification").submit();
+	$("#frm-UpdateUCertification").attr("action", "dashboard/" + currentItem + "/updateUCertification.jv");
+	$("#frm-UpdateUCertification").submit();
 }
 
 function deleteUCertification() {
-	window.location.href="dashboard/" + currentItem + "/deleteUCertification.jv";
+	var isDelete = confirm("Do you want to delete this Certification?");
+	if(isDelete) {
+		window.location.href="dashboard/" + currentItem + "/deleteUCertification.jv";
+	}
+}
+
+/**bind current License item to edit form*/
+function bindCurrentLicense(licenseId){
+	currentItem = licenseId;
+	
+	$("#license-conferringOrganization").val($("#license-conferringOrganization-"+ currentItem+" h5").text());
+	
+	$("#license-professionalLicense").val($("#license-professionalLicense-"+ currentItem).text());
+	
+	$("#license-dateIssued").val($("#license-dateIssued-"+ currentItem).text());
+	
+	$("#license-licenseNumber").val($("#license-licenseNumber-"+ currentItem).text());
+	
+	$("#license-description").val($("#license-description-"+ currentItem).text());
+}
+
+function updateULicense() {
+	$("#frm-UpdateULicense").attr("action", "dashboard/" + currentItem + "/updateULicense.jv");
+	$("#frm-UpdateULicense").submit();
+}
+
+function deleteULicense() {
+	var isDelete = confirm("Do you want to delete this License?");
+	if(isDelete) {
+		window.location.href="dashboard/" + currentItem + "/deleteULicense.jv";
+	}
+}
+
+/**bind current Education item to edit form*/
+function bindCurrentEducation(educationId){
+	currentItem = educationId;
+	
+	$("#education-institutionName").val($("#education-institutionName-"+ currentItem+ " a h5").text());
+	
+	$("#education-degreeType").val($("#education-degreeType-"+ currentItem).text());
+	
+	$("#education-graduationStartDate").val($("#education-graduationStartDate-"+ currentItem).text());
+	
+	$("#education-graduationEndDate").val($("#education-graduationEndDate-"+ currentItem).text());
+	
+	$("#education-description").val($("#education-description-"+ currentItem).text());
+}
+
+function updateUEducation() {
+	$("#frm-UpdateUEducation").attr("action", "dashboard/" + currentItem + "/updateUEducation.jv");
+	$("#frm-UpdateUEducation").submit();
+}
+
+function deleteUEducation() {
+	var isDelete = confirm("Do you want to delete this Education?");
+	if(isDelete) {
+		window.location.href="dashboard/" + currentItem + "/deleteUEducation.jv";
+	}
+}
+
+/**bind current Employment item to edit form*/
+function bindCurrentEmployment(employmentId){
+	currentItem = employmentId;
+	
+	$("#employment-clientName").val($("#employment-clientName-"+ currentItem+ " a h5").text());
+	
+	$("#employment-positionHeld").val($("#employment-positionHeld-"+ currentItem).text());
+	
+	$("#employment-startDate").val($("#employment-startDate-"+ currentItem).text());
+	
+	$("#employment-endDate").val($("#employment-endDate-"+ currentItem).text());
+	
+	$("#employment-description").val($("#employment-description-"+ currentItem).text());
+}
+
+function updateUEmployment() {
+	$("#frm-UpdateUEmployment").attr("action", "dashboard/" + currentItem + "/updateUEmployment.jv");
+	$("#frm-UpdateUEmployment").submit();
+}
+
+function deleteUEmployment() {
+	var isDelete = confirm("Do you want to delete this Employment?");
+	if(isDelete) {
+		window.location.href="dashboard/" + currentItem + "/deleteUEmployment.jv";
+	}
 }
